Resolved an issue with Linux platform.
