+++ .github/workflows/deploy.yml
@@ -22,8 +22,8 @@ jobs:
         include:
           - target: publishIosArm64PublicationToSonatypeRepository
             os: macos-latest
-#          - target: publishLinuxX64PublicationToSonatypeRepository
-#            os: ubuntu-latest
+          - target: publishLinuxX64PublicationToSonatypeRepository
+            os: ubuntu-latest
           - target: publishJvmPublicationToSonatypeRepository
             os: ubuntu-latest
           - target: publishKotlinMultiplatformPublicationToSonatypeRepository
