+++ cli/src/nativeMain/kotlin/com/kss/zoom/cli/NativeTerminalManager.kt
@@ -1,24 +1,13 @@
 package com.kss.zoom.cli
 
 import com.github.ajalt.mordant.terminal.Terminal
-import kotlinx.cinterop.*
-import platform.posix.*
+import kotlinx.cinterop.refTo
+import platform.posix.STDIN_FILENO
+import platform.posix.read
 
-class NativeTerminalManager(private val terminal: Terminal) : TerminalManager {
+abstract class NativeTerminalManager(private val terminal: Terminal) : TerminalManager {
 
     private val inputListeners = mutableListOf<(Char) -> Unit>()
-
-    override fun enableRawMode() {
-        memScoped {
-            val term = alloc<termios>()
-            tcgetattr(STDIN_FILENO, term.ptr)
-
-            term.c_lflag = term.c_lflag and ICANON.inv().toULong() // disable canonical mode and echo
-            term.c_cc[VMIN] = 1u // minimum number of characters for noncanonical read
-            term.c_cc[VTIME] = 0u // timeout for noncanonical read
-            tcsetattr(STDIN_FILENO, TCSANOW, term.ptr)
-        }
-    }
 
     override fun captureInput(): String {
         val buffer = ByteArray(1)
