+++ sdk/build.gradle.kts
@@ -14,7 +14,7 @@ jacoco {
 
 val nativeTarget = when (val hostOs = System.getProperty("os.name")) {
     "Mac OS X" -> "MacosArm64"
-//    "Linux" -> "LinuxX64"
+    "Linux" -> "LinuxX64"
     else -> throw GradleException("Host $hostOs is not supported in Kotlin/Native.")
 }
 
@@ -23,8 +23,7 @@ fun KotlinNativeTargetWithHostTests.configureTarget() =
 
 kotlin {
     macosArm64 { configureTarget() }
-    // See the client -> Blocked by an issue with curl: https://youtrack.jetbrains.com/issue/KTOR-6361
-//    linuxX64 { configureTarget() }
+    linuxX64 { configureTarget() }
 
     jvm {
         compilations.all {
@@ -126,11 +125,11 @@ kotlin {
         val posixTest by creating {
             dependsOn(nativeTest)
         }
-        arrayOf("macosArm64" /*, "linuxX64"*/).forEach { targetName ->
+        arrayOf("macosArm64" , "linuxX64").forEach { targetName ->
             getByName("${targetName}Main").dependsOn(posixMain)
             getByName("${targetName}Test").dependsOn(posixTest)
         }
-        arrayOf("macosArm64" /*, "linuxX64"*/).forEach { targetName ->
+        arrayOf("macosArm64" , "linuxX64").forEach { targetName ->
             getByName("${targetName}Main").dependsOn(nativeMain)
             getByName("${targetName}Test").dependsOn(nativeTest)
         }
